WebPlatform/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit-searchbox>=0.1.0
pydeck>=0.8.0
